+++ units/Entity.cpp
@@ -9,13 +9,13 @@ Entity::Entity()
 
 
 void Entity::update( const int& dt )
-{	    
+{
 	if( !isMoved() ){
-	    float x = getUnitX() + ( -150 + ( rand() % 300 ) ); 
+	    float x = getUnitX() + ( -150 + ( rand() % 300 ) );
 	    float y = getUnitY() - ( -150 + ( rand() % 300 ) );
 	    setPathTarget(x, y);
-	}	
-	move( dt );	
+	}
+	move( dt );
 }
 
 void Entity::move( const int& dt )
@@ -23,16 +23,16 @@ void Entity::move( const int& dt )
 	if( !isMoved() )
 	    return;
 	int dx = 0, dy = 0;
-	int nx = gotoX - getUnitX();
-	int ny = gotoY - getUnitY();
-	if( abs(nx) > 1 )	
+	int nx = static_cast<int>(gotoX - getUnitX());
+	int ny = static_cast<int>(gotoY - getUnitY());
+	if( abs(nx) > 1 )
 	    dx = nx / abs(nx);
 	if( abs(ny) > 1 )
 	    dy = ny / abs(ny);
 	if( dx == 0 && dy == 0){
 	    stopMove();
 	    return;
-	}	
+	}
 	moveUnit( dx, dy, dt );
 }
